1d8956a0 Add f2fs events [b/73625480]
5e38ea28 Remove inodes we find from scheduled scan as well [b/73625480]

The merge-list contains also some test refactorings and test cleanup CLs.
This CL is deliberately merging them as well even if they are not tied
to P1 bugs for the following reason:
1) Not doing so would break our treehugger tests on pi-dev
2) Trying to do manual surgery on the branch to cut them away is just
   going to increase the chance of screwing up future cherry-picks.
3) Test changes don't affect stability risk of the release.

Bug: 73625480
Test: manual + perfetto_unittests + perfetto_integrationtests

+++ include/perfetto/base/watchdog_posix.h
@@ -63,7 +63,7 @@ class Watchdog {
   // averaged over the last |window_ms| milliseconds. If |kb| is 0, any
   // existing limit is removed.
   // Note: |window_ms| has to be a multiple of |polling_interval_ms_|.
-  void SetMemoryLimit(uint32_t bytes, uint32_t window_ms);
+  void SetMemoryLimit(uint64_t bytes, uint32_t window_ms);
 
   // Sets a limit on the CPU usage used by the program averaged over the last
   // |window_ms| milliseconds. If |percentage| is 0, any existing limit is
@@ -139,7 +139,7 @@ class Watchdog {
   std::mutex mutex_;
   bool quit_ = true;
 
-  uint32_t memory_limit_bytes_ = 0;
+  uint64_t memory_limit_bytes_ = 0;
   WindowedInterval memory_window_bytes_;
 
   uint32_t cpu_limit_percentage_ = 0;
