+++ src/tracing/core/trace_packet_unittest.cc
@@ -132,7 +132,8 @@ TEST(TracePacketTest, GetProtoPreamble) {
   ASSERT_EQ(1u, tp.slices().size());
   memcpy(&buf[preamble_size], tp.slices()[0].start, tp.slices()[0].size);
   protos::Trace trace;
-  ASSERT_TRUE(trace.ParseFromArray(buf, preamble_size + tp.size()));
+  ASSERT_TRUE(
+      trace.ParseFromArray(buf, static_cast<int>(preamble_size + tp.size())));
   ASSERT_EQ(1, trace.packet_size());
   ASSERT_EQ(payload, trace.packet(0).for_testing().str());
 }
